Shiny infrastructure works but multiple tabs for client, proj, phase, task doesn't allow for plotting the same figure on different tabs (means that switch fcn can't be used in server file). This is the last commit with multiple tabs before trying to consolidate into a single tab with filters.
was having the problem of the file uploader not working if figures were output on the second tab. this is resolved when a *different* figure is output on the second tab - it isn't a plotly issue or an issue with the file or date inputs.

(this commit includes test code for histograms to show that everything works as expected).